create FilterContext in PageAdmin. FilterContainer moved in Filter.jsx (container and pure component in one file).
